Move process monitoring/cancelling functions from main app to utils file. For testing purposes: put TJs original launch_simulation code and cancel-button testing code in separate functions called within launch_simulation for easier visualization/testing.
