+++ app.py
@@ -45,6 +45,8 @@ from scripts.update_dir_list import UpdateDirList
 from scripts.update_hodo_page import UpdateHodoHTML
 from scripts.nse import Nse
 
+import utils 
+
 # Earth radius (km)
 R = 6_378_137
 
@@ -181,7 +183,9 @@ class RadarSimulator(Config):
         self.scripts_path = self.current_dir / 'scripts'
         self.obs_script_path = self.current_dir / 'obs_placefile.py'
         self.hodo_script_path = self.scripts_path / 'hodo_plot.py'
-        self.nexrad_script_path = self.scripts_path / 'get_nexrad.py'
+        self.nexrad_script_path = self.scripts_path / 'Nexrad.py'
+        self.l2munger_script_path = self.scripts_path / 'munger.py'
+        self.nse_script_path = self.scripts_path / 'nse.py'
         self.munge_dir = self.scripts_path / 'munge'
         self.assets_dir = self.current_dir / 'assets'
         self.hodo_images = self.assets_dir / 'hodographs'
@@ -568,18 +572,236 @@ def run_hodo_script(args) -> None:
     subprocess.run(["python", HODO_SCRIPT_PATH] + args, check=True)
 
 
-def exec_script(script_name, args):
+def run_TJ_original():
     """
-    Generalized function to run application scripts. subprocess.run() or similar is 
-    required for tracking of spawned python processes and termination if requested
-    by the user via the cancel button. Returns Exception object which is parsed to 
-    determine exit code/status. 
+    This is the "OG" script-launcher. 
     """
+    sa.scripts_progress = 'Setting up files and times'
+    # determine actual event time, playback time, diff of these two
+    sa.make_simulation_times()
+
+    # clean out old files and directories
     try:
-        subprocess.run(["python", f"{sa.scripts_path}/{script_name}"] + args, check=True)
+        sa.remove_files_and_dirs()
     except Exception as e:
-        print(f"Error running {script_name}: {e}")
-        return e
+        print("Error removing files and directories: ", e)
+
+    # based on list of selected radars, create a dictionary of radar metadata
+    try:
+        sa.create_radar_dict()
+        # sa.create_grlevel2_cfg_file()
+        sa.copy_grlevel2_cfg_file()
+    except Exception as e:
+        print("Error creating radar dict or config file: ", e)
+
+    # acquire radar data for the event
+    sa.scripts_progress = 'Downloading radar data ...'
+    try:
+        # Initial for loop to gather all radar files. Not great, but not sure of a better
+        # way to handle this. Calls NexradDownloader but passes download=False to only
+        # query AWS for expected files
+        query_radar_files()
+
+        for _r, radar in enumerate(sa.radar_list):
+            radar = radar.upper()
+            try:
+                if sa.new_radar == 'None':
+                    new_radar = radar
+                else:
+                    new_radar = sa.new_radar.upper()
+            except Exception as e:
+                print("Error defining new radar: ", e)
+            try:
+                print(
+                    f"Nexrad Downloader - {radar}, {sa.event_start_str}, {str(sa.event_duration)}")
+                _file_list = NexradDownloader(radar, sa.event_start_str, str(sa.event_duration),
+                                                download=True)
+                # sa.radar_dict[radar]['file_list'] = file_list
+                time.sleep(10)
+            except Exception as e:
+                print("Error running nexrad script: ", e)
+            try:
+                print(f"Munge from {radar} to {new_radar}...")
+                Munger(radar, sa.playback_start_str, sa.event_duration, sa.simulation_seconds_shift,
+                        new_radar, playback_speed=1.5)
+                print(f"Munge for {new_radar} completed ...")
+                # this gives the user some radar data to poll while other scripts are running
+                UpdateDirList(new_radar, 'None', initialize=True)
+
+            except Exception as e:
+                print(
+                    "Error running Munge from {radar} to {new_radar}: ", e)
+    except Exception as e:
+        print("Error running nexrad or munge scripts: ", e)
+
+    sa.scripts_progress = 'Creating obs placefiles ...'
+    try:
+        print("Running obs script...")
+        Mesowest(str(sa.lat), str(sa.lon),
+                    sa.event_start_str, str(sa.event_duration))
+        print("Obs script completed")
+    except Exception as e:
+        print("Error running obs script: ", e)
+
+    sa.scripts_progress = 'Creating NSE placefiles ...'
+    # NSE placefiles
+    try:
+        print("Running NSE scripts...")
+        Nse(str(sa.event_start_time), sa.event_duration, sa.scripts_path, sa.data_dir,
+            sa.placefiles_dir)
+    except Exception as e:
+        print("Error running NSE scripts: ", e)
+
+    # Since there will always be a timeshift associated with a simulation, this
+    # script needs to execute every time, even if a user doesn't select a radar
+    # to transpose to.
+    run_transpose_script()
+    # sa.rename_shifted_nse_placefiles()
+
+    sa.scripts_progress = 'Creating hodo plots ...'
+
+    for radar, data in sa.radar_dict.items():
+        try:
+            asos_one = data['asos_one']
+            asos_two = data['asos_two']
+            print(asos_one, asos_two, radar)
+        except KeyError as e:
+            print("Error getting radar metadata: ", e)
+
+        try:
+            info = f'hodo script:{radar},{sa.new_radar},{asos_one},{asos_two},{sa.simulation_seconds_shift}'
+            print(info)
+            run_hodo_script([radar, sa.new_radar, asos_one,
+                            asos_two, str(sa.simulation_seconds_shift)])
+            print("Hodograph script completed ...")
+        except Exception as e:
+            print("Error running hodo script: ", e)
+        try:
+            UpdateHodoHTML('None', initialize=True)
+        except Exception as e:
+            print("Error updating hodo html: ", e)
+
+    sa.scripts_progress = 'Scripts completed!'
+
+
+def run_with_cancel_button():
+    """
+    This version of the script-launcher trying to work in cancel button
+    """
+    sa.scripts_progress = 'Setting up files and times'
+    # determine actual event time, playback time, diff of these two
+    sa.make_simulation_times()
+
+    # clean out old files and directories
+    try:
+        sa.remove_files_and_dirs()
+    except Exception as e:
+        print("Error removing files and directories: ", e)
+
+    # based on list of selected radars, create a dictionary of radar metadata
+    try:
+        sa.create_radar_dict()
+        # sa.create_grlevel2_cfg_file()
+        sa.copy_grlevel2_cfg_file()
+    except Exception as e:
+        print("Error creating radar dict or config file: ", e)
+
+    # acquire radar data for the event
+    sa.scripts_progress = 'Downloading radar data ...'
+    try:
+        # Initial for loop to gather all radar files. Not great, but not sure of a better
+        # way to handle this. Calls NexradDownloader but passes download=False to only
+        # query AWS for expected files
+        query_radar_files()
+
+        for _r, radar in enumerate(sa.radar_list):
+            radar = radar.upper()
+            try:
+                if sa.new_radar == 'None':
+                    new_radar = radar
+                else:
+                    new_radar = sa.new_radar.upper()
+            except Exception as e:
+                print("Error defining new radar: ", e)
+            
+            # Radar download
+            print(
+                f"Nexrad Downloader - {radar}, {sa.event_start_str}, {str(sa.event_duration)}"
+            )
+            args = [radar, f'{sa.event_start_str}', str(sa.event_duration), str(True)]
+            e = utils.exec_script(sa.nexrad_script_path, args)
+            # This section is what forces the callback to end if the cancel button was hit.
+            # The returncode for the exception equals the SIGTERM value (usually 15).
+            if e and e.returncode == -1*signal.SIGTERM:
+                return
+            
+            # Munger
+            print(f"Munge from {radar} to {new_radar}...")
+            args = [radar, str(sa.playback_start_str), str(sa.event_duration), str(sa.simulation_seconds_shift),
+                    new_radar]
+            e = utils.exec_script(sa.l2munger_script_path, args)
+            if e and e.returncode == -1*signal.SIGTERM:
+                return
+            print(f"Munge for {new_radar} completed ...")
+            
+            # this gives the user some radar data to poll while other scripts are running
+            try:
+                UpdateDirList(new_radar, 'None', initialize=True)
+            except Exception as e:
+                print(f"Error with UpdateDirList ", e)
+                            
+    except Exception as e:
+        print("Error running nexrad or munge scripts: ", e)
+
+    sa.scripts_progress = 'Creating obs placefiles ...'
+    print("Running obs script...")
+    args = [str(sa.lat), str(sa.lon), sa.event_start_str, str(sa.event_duration)]
+    e = utils.exec_script(sa.obs_script_path, args)
+    if e and e.returncode == -1*signal.SIGTERM:
+        return
+
+    sa.scripts_progress = 'Creating NSE placefiles ...'
+    print("Running NSE scripts...")
+    args = [str(sa.event_start_time), str(sa.event_duration), str(sa.scripts_path), 
+            str(sa.data_dir), str(sa.placefiles_dir)]
+    #e = utils.exec_script(sa.nse_script_path, args)
+    if e and e.returncode == -1*signal.SIGTERM:
+        return
+    
+    # Since there will always be a timeshift associated with a simulation, this
+    # script needs to execute every time, even if a user doesn't select a radar
+    # to transpose to.
+    run_transpose_script()
+    # sa.rename_shifted_nse_placefiles()
+
+    # Hodographs 
+    sa.scripts_progress = 'Creating hodo plots ...'
+    for radar, data in sa.radar_dict.items():
+        try:
+            asos_one = data['asos_one']
+            asos_two = data['asos_two']
+            print(asos_one, asos_two, radar)
+        except KeyError as e:
+            print("Error getting radar metadata: ", e)
+        
+        info = f'hodo script:{radar},{sa.new_radar},{asos_one},{asos_two},{sa.simulation_seconds_shift}'
+        print(info)
+        
+        # Execute hodograph script
+        args = [radar, sa.new_radar, asos_one, asos_two, 
+                str(sa.simulation_seconds_shift)]
+        e = utils.exec_script(sa.hodo_script_path, args)
+        if e and e.returncode == -1*signal.SIGTERM:
+            return
+        print("Hodograph script completed ...")
+
+        try:
+            UpdateHodoHTML('None', initialize=True)
+        except Exception as e:
+            print("Error updating hodo html: ", e)
+    
+        sa.scripts_progress = 'Scripts completed!'
+
 
 @app.callback(
     Output('show_script_progress', 'children'),
@@ -607,301 +829,22 @@ def launch_simulation(n_clicks) -> None:
     if n_clicks == 0:
         raise PreventUpdate
     else:
-        sa.scripts_progress = 'Setting up files and times'
-        # determine actual event time, playback time, diff of these two
-        sa.make_simulation_times()
-
-        # clean out old files and directories
-        try:
-            sa.remove_files_and_dirs()
-        except Exception as e:
-            print("Error removing files and directories: ", e)
-
-        # based on list of selected radars, create a dictionary of radar metadata
-        try:
-            sa.create_radar_dict()
-            # sa.create_grlevel2_cfg_file()
-            sa.copy_grlevel2_cfg_file()
-        except Exception as e:
-            print("Error creating radar dict or config file: ", e)
-
-        # acquire radar data for the event
-        sa.scripts_progress = 'Downloading radar data ...'
-        try:
-            # Initial for loop to gather all radar files. Not great, but not sure of a better
-            # way to handle this. Calls NexradDownloader but passes download=False to only
-            # query AWS for expected files
-            query_radar_files()
-
-            for _r, radar in enumerate(sa.radar_list):
-                radar = radar.upper()
-                try:
-                    if sa.new_radar == 'None':
-                        new_radar = radar
-                    else:
-                        new_radar = sa.new_radar.upper()
-                except Exception as e:
-                    print("Error defining new radar: ", e)
-                
-                # Radar download
-                print(
-                    f"Nexrad Downloader - {radar}, {sa.event_start_str}, {str(sa.event_duration)}"
-                )
-                args = [radar, f'{sa.event_start_str}', str(sa.event_duration), str(True)]
-                e = exec_script("Nexrad.py", args)
-                # This section is what forces the callback to end if the cancel button was hit.
-                # The returncode for the exception equals the SIGTERM value (usually 15).
-                if e and e.returncode == -1*signal.SIGTERM:
-                    return
-                
-                # Munger
-                print(f"Munge from {radar} to {new_radar}...")
-                args = [radar, str(sa.playback_start_str), str(sa.event_duration), str(sa.simulation_seconds_shift),
-                        new_radar]
-                e = exec_script("munger.py", args)
-                if e and e.returncode == -1*signal.SIGTERM:
-                    return
-                print(f"Munge for {new_radar} completed ...")
-                
-                # this gives the user some radar data to poll while other scripts are running
-                try:
-                    UpdateDirList(new_radar, 'None', initialize=True)
-                except Exception as e:
-                    print(f"Error with UpdateDirList ", e)
-                                
-        except Exception as e:
-            print("Error running nexrad or munge scripts: ", e)
-
-        sa.scripts_progress = 'Creating obs placefiles ...'
-        print("Running obs script...")
-        args = [str(sa.lat), str(sa.lon), sa.event_start_str, str(sa.event_duration)]
-        e = exec_script("obs_placefile.py", args)
-        if e and e.returncode == -1*signal.SIGTERM:
-            return
-
-        sa.scripts_progress = 'Creating NSE placefiles ...'
-        print("Running NSE scripts...")
-        args = [str(sa.event_start_time), str(sa.event_duration), str(sa.scripts_path), 
-                str(sa.data_dir), str(sa.placefiles_dir)]
-        e = exec_script("nse.py", args)
-        if e and e.returncode == -1*signal.SIGTERM:
-            return
+        #run_with_cancel_button()
+        run_TJ_original()
         
-        # Since there will always be a timeshift associated with a simulation, this
-        # script needs to execute every time, even if a user doesn't select a radar
-        # to transpose to.
-        run_transpose_script()
-        # sa.rename_shifted_nse_placefiles()
 
-        # Hodographs 
-        sa.scripts_progress = 'Creating hodo plots ...'
-        for radar, data in sa.radar_dict.items():
-            try:
-                asos_one = data['asos_one']
-                asos_two = data['asos_two']
-                print(asos_one, asos_two, radar)
-            except KeyError as e:
-                print("Error getting radar metadata: ", e)
-            
-            info = f'hodo script:{radar},{sa.new_radar},{asos_one},{asos_two},{sa.simulation_seconds_shift}'
-            print(info)
-            
-            # Execute hodograph script
-            args = [radar, sa.new_radar, asos_one, asos_two, 
-                    str(sa.simulation_seconds_shift)]
-            e = exec_script("hodo_plot.py", args)
-            if e and e.returncode == -1*signal.SIGTERM:
-                return
-            print("Hodograph script completed ...")
-
-            try:
-                UpdateHodoHTML('None', initialize=True)
-            except Exception as e:
-                print("Error updating hodo html: ", e)
-        
-        sa.scripts_progress = 'Scripts completed!'
-
-'''
-# Original launch_simulation function
-@app.callback(
-    Output('show_script_progress', 'children'),
-    [Input('run_scripts', 'n_clicks')],
-    prevent_initial_call=True,
-    running=[
-        (Output('start_year', 'disabled'), True, False),
-        (Output('start_month', 'disabled'), True, False),
-        (Output('start_day', 'disabled'), True, False),
-        (Output('start_hour', 'disabled'), True, False),
-        (Output('start_minute', 'disabled'), True, False),
-        (Output('duration', 'disabled'), True, False),
-        (Output('radar_quantity', 'disabled'), True, False),
-        (Output('map_btn', 'disabled'), True, False),
-        (Output('new_radar_selection', 'disabled'), True, False),
-        (Output('run_scripts', 'disabled'), True, False),
-        (Output('playback-clock', 'disabled'), True, False),
-    ])
-def launch_simulation(n_clicks) -> None:
-    """
-    This function is called when the "Run Scripts" button is clicked. It will execute the
-    necessary scripts to simulate radar operations, create hodographs, and transpose placefiles.
-    """
-    if n_clicks == 0:
-        raise PreventUpdate
-    else:
-        sa.scripts_progress = 'Setting up files and times'
-        # determine actual event time, playback time, diff of these two
-        sa.make_simulation_times()
-
-        # clean out old files and directories
-        try:
-            sa.remove_files_and_dirs()
-        except Exception as e:
-            print("Error removing files and directories: ", e)
-
-        # based on list of selected radars, create a dictionary of radar metadata
-        try:
-            sa.create_radar_dict()
-            # sa.create_grlevel2_cfg_file()
-            sa.copy_grlevel2_cfg_file()
-        except Exception as e:
-            print("Error creating radar dict or config file: ", e)
-
-        # acquire radar data for the event
-        sa.scripts_progress = 'Downloading radar data ...'
-        try:
-            # Initial for loop to gather all radar files. Not great, but not sure of a better
-            # way to handle this. Calls NexradDownloader but passes download=False to only
-            # query AWS for expected files
-            query_radar_files()
-
-            for _r, radar in enumerate(sa.radar_list):
-                radar = radar.upper()
-                try:
-                    if sa.new_radar == 'None':
-                        new_radar = radar
-                    else:
-                        new_radar = sa.new_radar.upper()
-                except Exception as e:
-                    print("Error defining new radar: ", e)
-                try:
-                    print(
-                        f"Nexrad Downloader - {radar}, {sa.event_start_str}, {str(sa.event_duration)}")
-                    _file_list = NexradDownloader(radar, sa.event_start_str, str(sa.event_duration),
-                                                  download=True)
-                    # sa.radar_dict[radar]['file_list'] = file_list
-                    time.sleep(10)
-                except Exception as e:
-                    print("Error running nexrad script: ", e)
-                try:
-                    print(f"Munge from {radar} to {new_radar}...")
-                    Munger(radar, sa.playback_start_str, sa.event_duration, sa.simulation_seconds_shift,
-                           new_radar, playback_speed=1.5)
-                    print(f"Munge for {new_radar} completed ...")
-                    # this gives the user some radar data to poll while other scripts are running
-                    UpdateDirList(new_radar, 'None', initialize=True)
-
-                except Exception as e:
-                    print(
-                        "Error running Munge from {radar} to {new_radar}: ", e)
-        except Exception as e:
-            print("Error running nexrad or munge scripts: ", e)
-
-        sa.scripts_progress = 'Creating obs placefiles ...'
-        try:
-            print("Running obs script...")
-            Mesowest(str(sa.lat), str(sa.lon),
-                     sa.event_start_str, str(sa.event_duration))
-            print("Obs script completed")
-        except Exception as e:
-            print("Error running obs script: ", e)
-
-        sa.scripts_progress = 'Creating NSE placefiles ...'
-        # NSE placefiles
-        try:
-            print("Running NSE scripts...")
-            Nse(sa.event_start_time, sa.event_duration, sa.scripts_path, sa.data_dir,
-                sa.placefiles_dir)
-        except Exception as e:
-            print("Error running NSE scripts: ", e)
-
-        # Since there will always be a timeshift associated with a simulation, this
-        # script needs to execute every time, even if a user doesn't select a radar
-        # to transpose to.
-        run_transpose_script()
-        # sa.rename_shifted_nse_placefiles()
-
-        sa.scripts_progress = 'Creating hodo plots ...'
-
-        for radar, data in sa.radar_dict.items():
-            try:
-                asos_one = data['asos_one']
-                asos_two = data['asos_two']
-                print(asos_one, asos_two, radar)
-            except KeyError as e:
-                print("Error getting radar metadata: ", e)
-
-            try:
-                info = f'hodo script:{radar},{sa.new_radar},{asos_one},{asos_two},{sa.simulation_seconds_shift}'
-                print(info)
-                run_hodo_script([radar, sa.new_radar, asos_one,
-                                asos_two, str(sa.simulation_seconds_shift)])
-                print("Hodograph script completed ...")
-            except Exception as e:
-                print("Error running hodo script: ", e)
-            try:
-                UpdateHodoHTML('None', initialize=True)
-            except Exception as e:
-                print("Error updating hodo html: ", e)
-
-        sa.scripts_progress = 'Scripts completed!'
-'''
 ################################################################################################
 # ----------------------------- Monitoring and reporting script status  ------------------------
 ################################################################################################
-
-
-def get_python_processes():
-    """
-    Reports back all running python processes as a list. Used by the monitoring 
-    function and cancel button.
-    """
-    processes = []
-    for proc in psutil.process_iter(['pid', 'cmdline', 'name', 'username']):
-        try:
-            info = proc.info
-            if 'python' in info['name'] and len(info['cmdline']) > 1: 
-                processes.append(info)
-        except:
-            pass
-    return processes 
 
 
 @app.callback(
     Output('dummy', 'data'),
     [Input('cancel_scripts', 'n_clicks')],   
     prevent_initial_call=True)
-def cancel_all(n_clicks):
+def cancel_scripts(n_clicks):
     if n_clicks > 0:
-        # Should move this somewhere else, maybe into the __init__ function? These are 
-        # the cancelable scripts
-        scripts_list = ["Nexrad.py", "nse.py", "get_data.py", "process.py", 
-                        "hodo_plot.py", "munger.py"]
-        processes = get_python_processes()
-
-        # ******************************************************************************
-        # POTENTIAL ISSUES - Race Conditions?
-        # There is a chance a new process could spawn between the initial query above
-        # and process(es) being terminated below. Also, a process could end before 
-        # getting into the loop. In that case, is it gauranteed that "old" pid isn't 
-        # re-used somewhere else, and we end up accidentally terminating something we 
-        # shouldn't be?
-        # ******************************************************************************
-        for process in processes:
-            if any(x in process['cmdline'][1] for x in scripts_list):
-                print(f"Killing process: {process['cmdline'][1]} with pid: {process['pid']}")
-                os.kill(process['pid'], signal.SIGTERM)
-                #p = psutil.Process(process['pid'])
-                #p.terminate()
+        utils.cancel_all()
 
 
 @app.callback(
@@ -916,7 +859,7 @@ def monitor(_n):
     status of the radar and hodograph scripts and reports the status to the user.
     """
     # Radar file download status
-    radar_dl_completion, radar_files = radar_monitor()
+    radar_dl_completion, radar_files = utils.radar_monitor(sa)
 
     # Hodographs. Currently hard-coded to expect 2 files for every radar and radar file.
     num_hodograph_images = len(glob(f"{sa.hodo_images}/*.png"))
@@ -926,58 +869,9 @@ def monitor(_n):
             (num_hodograph_images / (2*len(radar_files)))
 
     # NSE placefiles
-    model_list = nse_status_checker()
+    model_list = utils.nse_status_checker(sa)
     return radar_dl_completion, hodograph_completion, model_list
 
-
-def radar_monitor():
-    """Reads radar_dict.json file(s) output from the NexradDownloader. Looks for associated 
-    radar files on the system and compares to the total expected number and broadcasts a 
-    percentage to the radar_status progress bar."""
-    expected_files = []
-    radar_dirs = glob(f"{sa.data_dir}/radar/**")
-    for d in radar_dirs:
-        json_file = glob(f"{d}/downloads/radar_dict.json")
-        if len(json_file) == 1:
-            with open(json_file[0], 'r', encoding='utf-8') as f:
-                radar_dictionary = json.load(f)
-                expected_files.extend(list(radar_dictionary.values()))
-
-    files_on_system = [x for x in expected_files if os.path.exists(x)]
-    output = 0
-    if len(expected_files) > 0:
-        output = 100 * (len(files_on_system) / len(expected_files))
-    return output, files_on_system
-
-
-def nse_status_checker():
-    """
-    Read in model status text file and query associated file sizes. 
-    """
-    filename = f"{sa.data_dir}/model_data/model_list.txt"
-    output = []
-    if os.path.exists(filename):
-        model_list = []
-        filesizes = []
-        with open(filename, 'r', encoding='utf-8') as f:
-            text_listing = f.readlines()
-            for line in text_listing:
-                filename = line[:-1]
-                model_list.append(filename.rsplit('/', 1)[1])
-                filesizes.append(round(file_stats(filename), 2))
-
-        df = pd.DataFrame({'Model data': model_list, 'Size (MB)': filesizes})
-        output = df.to_dict('records')
-
-    return output
-
-
-def file_stats(filename):
-    """Return the size of a specific file.  If it doesn't exist, returns 0"""
-    filesize = 0.
-    if os.path.exists(filename):
-        filesize = os.stat(filename).st_size / 1024000.
-    return filesize
 
 # -------------------------------------
 # --- Transpose placefiles in time and space
